General Modifications REST way
